moved inspectors to ruleset-respective folders
